+++ packages/common/.eslintignore
@@ -1,0 +1,3 @@
+/.eslintrc.js
+/dist
+/build.mjs
