+++ packages/common/package.json
@@ -1,15 +1,14 @@
 {
-  "name": "@preevy/compose-tunnel-agent",
+  "name": "@preevy/common",
   "version": "0.0.16",
   "description": "",
   "main": "dist/index.js",
   "files": [
     "dist",
-    "Dockerfile"
+    "tsconfig.json"
   ],
   "license": "Apache-2.0",
   "dependencies": {
-    "@preevy/common": "0.0.16",
     "dockerode": "^3.3.4",
     "http-proxy": "^1.18.1",
     "lodash": "^4.17.21",
@@ -34,12 +33,9 @@
     "typescript": "^5.0.2"
   },
   "scripts": {
-    "start": "node dist/index.js",
-    "dev": "tsx watch ./index.ts",
+    "start": "node dist/main.js",
+    "dev": "tsx watch ./main.ts",
     "lint": "eslint . --ext .ts,.tsx",
-    "clean": "rm -rf dist",
-    "build": "node build.mjs",
-    "prepack": "yarn build",
-    "prepare": "cd ../.. && husky install"
+    "build": "tsc -b"
   }
 }
