+++ packages/common/src/emitter/simple.ts
