+++ packages/common/src/env.ts
@@ -5,3 +5,8 @@ export const requiredEnv = (key: string): string => {
   }
   return result
 }
+
+export const numberFromEnv = (key: string) => {
+  const s = process.env[key]
+  return s === undefined ? undefined : Number(s)
+}
