+++ packages/common/src/ssh/connection-checker.ts
