+++ packages/tunnel-server/index.ts
@@ -1,6 +1,6 @@
 import { promisify } from 'util'
+import { requiredEnv, numberFromEnv } from '@preevy/common'
 import { app as createApp } from './src/app'
-import { numberFromEnv, requiredEnv } from './src/env'
 import { inMemoryPreviewEnvStore } from './src/preview-env'
 import { sshServer as createSshServer } from './src/ssh-server'
 import { getSSHKeys } from './src/ssh-keys'
