+++ packages/tunnel-server/package.json
@@ -6,6 +6,7 @@
   "license": "Apache-2.0",
   "dependencies": {
     "@fastify/request-context": "^4.2.0",
+    "@preevy/common": "0.0.16",
     "@sinclair/typebox": "^0.25.21",
     "fastify": "^4.12.0",
     "http-proxy": "^1.18.1",
